utf-8에서 utf-8-bom으로 변경해서 경고 삭제. 동일 크기의 메모리를 관리해주는 메모리 풀 구현. 레드블랙트리 검색만 구현.

+++ Floater/FloaterRendererDX11/DX11Mesh.h
@@ -48,7 +48,7 @@ namespace flt
 		UINT singleVertexSize;
 
 		ID3D11Buffer* indexBuffer;
-		size_t indexCount;
+		UINT indexCount;
 
 		ID3D11ShaderResourceView* texture;
 		ID3D11SamplerState* sampler;
